Add starter code for assignment
This change comprises starter code for the assignment -- e.g., the
declaration of the arithmetic functions that must be implemented and
stubs for those functions.

Unit tests for the arithmetic functions are also included. The unit
tests use munit, which is simple to integrate and supports
reproducible pseudo-random number generation. In addition, stylstic
checks are performed using cpplint and Cppcheck.

References:

- cppcheck: http://cppcheck.sourceforge.io/
- cpplint: https://github.com/cpplint/cpplint
- munit: https://nemequ.github.io/munit/